start on-disk backing for memory cache
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@17726 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ willow/src/include/flowio.h
@@ -306,7 +306,7 @@ struct buffering_filter : sink, spigot {
 	sink_result data_ready(char const *buf, size_t len, ssize_t &discard) {
 	sink_result	result;
 	ssize_t		d = discard;
-		while ((discard - d) < len) {
+		while ((size_t)(discard - d) < len) {
 			result = bf_transform(buf + (discard - d), len - (discard - d), discard);
 			if (result != sink_result_okay &&
 			    result != sink_result_finished)
